+++ include/api/CDataFrameTrainBoostedTreeRegressionRunner.h
@@ -56,9 +56,6 @@ public:
     //! \return A serialisable metadata of the trained regression model.
     TOptionalInferenceModelMetadata inferenceModelMetadata() const override;
 
-    //! \return A serialisable data summarization for the trained regression model
-    TDataSummarizationUPtr dataSummarization(const core::CDataFrame& dataFrame) const override;
-
 private:
     void validate(const core::CDataFrame& frame,
                   std::size_t dependentVariableColumn) const override;
