+++ include/api/CInferenceModelDefinition.h
@@ -6,14 +6,10 @@
 #ifndef INCLUDED_ml_api_CInferenceModelDefinition_h
 #define INCLUDED_ml_api_CInferenceModelDefinition_h
 
-#include <core/CRapidJsonConcurrentLineWriter.h>
-
 #include <maths/CDataFrameCategoryEncoder.h>
 
+#include <api/CSerializableToJson.h>
 #include <api/ImportExport.h>
-
-#include <rapidjson/document.h>
-#include <rapidjson/ostreamwrapper.h>
 
 #include <boost/optional.hpp>
 #include <boost/unordered_map.hpp>
@@ -24,39 +20,6 @@
 
 namespace ml {
 namespace api {
-
-//! \brief Abstract class for all elements the the inference definition
-//! that can will be serialized into JSON.
-class API_EXPORT CSerializableToJsonDocument {
-public:
-    using TRapidJsonWriter = core::CRapidJsonConcurrentLineWriter;
-
-public:
-    virtual ~CSerializableToJsonDocument() = default;
-    //! Serialize the object as JSON items under the \p parentObject using the specified \p writer.
-    virtual void addToJsonDocument(rapidjson::Value& parentObject,
-                                   TRapidJsonWriter& writer) const = 0;
-};
-
-class API_EXPORT CSerializableToJsonStream {
-public:
-    using TGenericLineWriter = core::CRapidJsonLineWriter<rapidjson::OStreamWrapper>;
-
-public:
-    virtual ~CSerializableToJsonStream() = default;
-    virtual void addToJsonStream(TGenericLineWriter& /*writer*/) const = 0;
-};
-
-class API_EXPORT CSerializableToJsonDocumentCompressed : public CSerializableToJsonStream{
-    public:
-        using TRapidJsonWriter = core::CRapidJsonConcurrentLineWriter;
-    public:
-    virtual ~CSerializableToJsonDocumentCompressed() = default;
-    virtual void addToDocumentCompressed(TRapidJsonWriter& writer) const = 0;
-    virtual void addToDocumentCompressed(TRapidJsonWriter& writer, std::string compressed_doc_tag) const;
-    virtual std::stringstream jsonCompressedStream() const;
-    virtual void jsonStream(std::ostream& jsonStrm) const;
-};
 
 //! Abstract class for output aggregation.
 class API_EXPORT CAggregateOutput : public CSerializableToJsonStream {
