+++ include/maths/CBoostedTree.h
@@ -242,7 +242,7 @@ public:
     CTreeShapFeatureImportance* shap() const override;
 
     //! Get the selected rows that summarize \p dataFrame.
-    core::CPackedBitVector dataSummarization(const core::CDataFrame& dataFrame) const;
+    core::CPackedBitVector dataSummarization(const core::CDataFrame& dataFrame) const override;
 
     //! Get the vector of hyperparameter importances.
     THyperparameterImportanceVec hyperparameterImportance() const;
