+++ include/maths/CDataFramePredictiveModel.h
@@ -66,6 +66,10 @@ public:
     //! \warning Will return a nullptr if a trained model isn't available.
     virtual CTreeShapFeatureImportance* shap() const = 0;
 
+    //! Get the selected rows that summarize \p dataFrame.
+    virtual core::CPackedBitVector
+    dataSummarization(const core::CDataFrame& dataFrame) const = 0;
+
     //! Get the column containing the dependent variable.
     virtual std::size_t columnHoldingDependentVariable() const = 0;
